Fixed Crash on Remove Attachment Button
The remove attachment function was referencing "self" when it called the update_field function. The update_field function is actually in the the NightLetterMain class, which is instantiated as "application" when the program is run. This fix allows the program to remove attachments instead of crashing when the Remove Attachments button is pushed.