Fixed Crash When Loading Previous Record
Fixed two bugs related to crashes when loading previous records.

The first crash occured when the DB records skipped a Letter ID. The program woud attempt to load an empty record and crash. Now it detects when a query returns no results and decrements the LetterID down until it finds a valid record.

The second crash occured when the query failed due to a problem decoding the field contents to UTF-8. This is probably due to some rich formatting that did not translate well when the Access records were exported. Currently, the program detects this error and skips to the next readable record.

For future reference:
-The fields in the DB need to be purged of unreadable characters at some point.
-The program crashes if it tries to load a record with a LetterID of less than zero. The goToPreviousRecord function needs a stop to keep the LetterID positive, or even better a way to detect the oldest DB record and not try to load anything from before the oldest record.