Fix Crash the Oldest DB Record
Added a conditional to the goToPreviousRecord function to keep the program from trying to load a record that has a lower LetterID than anything in the db.

Added a formatting check that hides the "Previous" button when the oldest record is show. This should remove the temptation to attempt to access a record older than the first DB record. Added a second formatting check to show the "Previous" button on any other record.