Fixed Crash after Clicking Next Record Button
Modified the goToNextRecord function to have a similar structure to the goToPreviousRecord function. Fixed crashes when the LetterIDs in the db skip a number or the fomatting of db field cannot be decoded.

Fixed a type in the formatting check to show the "Previous" button if the oldest record is not being displayed.