Attached Files Now Stored in Folders Based on LetterID
Modified the attach and remove functions to store attachments in separate subfolders in the Attachment folder. This will simplify migration of attachments from the existing database as well as file naming criteria.

Removed the feature that adds the date to the file name, since the separate folders for each record reduce the likelihood of having two attachments in a folder with the same name.