Removed JPH Fields and Set Shift Codes to Update
Removed the fields for listing JPH in various areas, as these were not used in the old night letter. We could add something like them back later if there is a way to pull that data programmatically, but no one wants to manually enter this info for every shift.

The shift letter codes for Prod. and Maint. shifts had not been updating from the db. Changed the load_all_fields function so that it updates them now.