Modified Text Field Layout
Removed unused fields and changed the layout of the text fields to two columns to reduce scrolling. This required modification of the function that finds all the fields that need to be modified for saving/loading/etc. Currently the allWidgets function looks in 4 hard-coded layouts for widgets to list. At some point, it would be nice to have it search the whole window programmatically, instead of just the hard-coded areas.