Added Function to Format Links after Migration
The links to attachments were dumped from Access in a format that was difficult to use in sqlite. The updateAttachmentLinks function works through all the attachment fields in the DB. It separates the file names, wraps them in a link wrapper pointed to the attachment path, then replaces the old link with the new in the attachment field.

Tweaked the data cleaning macros to prepare the links for reformatting.